Add boolean variable for whether or not to associate a Load Balancer
Got bit by the terrible hashicorp/terraform#12570 which means that
`count` arguments can not use values from resources. This means we need
to create a boolean flag for turning this functionality on/off.

I also took the liberty of switch the "alb_" variables to "lb_" since
those are more correct (you can use an NLB) and it will force users to
add the boolean flag if they are using a Load Balancer.